jiong

+++ oj/models.py
@@ -39,7 +39,7 @@ class Solution(models.Model):
 	solution_id = models.AutoField(primary_key=True)
 	problem = models.ForeignKey(Problem)
 	user = models.ForeignKey(User)
-	score = models.IntegerField(default=0)
+	score = models.FloatField(default=0.0)
 	time = models.IntegerField(default=0)
 	memory = models.IntegerField(default=0) # ...
 	in_date = models.DateTimeField(auto_now_add=True)
